requirements.txt
requests>=2.31.0
finnhub-python>=2.4.18
beautifulsoup4>=4.12.2
readability-lxml>=0.8.1
playwright>=1.42.0
